backend/requirements.txt
fastapi
uvicorn
pydantic
sqlalchemy
anthropic
google-genai
python-dotenv
httpx
